Create SECURITY.md
Hello Core Team  Pi Network
I am one of the pioneers of your great project that serves all the people of the world, and I would like to thank you for this great effort in advancing the world of encryption.
Since I am a member of the Pi Network community, I want to deliver serious information to the RafiosForUs project. When logging in to pay on the Test Mainnet, I discovered a dangerous feature that requests 24 passwords for the Pi Browser wallet. This request may lead to the theft of the Pi Browser wallets to the Pi Network community.
Accept my full respect and appreciation